fix(frontend): clear react-doctor warnings on the promotion diff
Ten React Doctor warnings across five files, all introduced by work already on
dev and surfaced against main by the promotion PR.

- ModalBase: read isTopmostModal through a ref, matching the closeModalRef
  idiom already in the file, so the document-level Escape and outside-click
  listeners attach once per open rather than re-subscribing on an identity
  change. The callback is stable today; the ref keeps that true if it ever
  gains dependencies.
- WalletPassPreview: nextDueLine walked the vaccination list five times
  (map, filter, map, filter, sort) to read a single earliest entry. Replaced
  with one pass tracking a running minimum.
- PublicPassportView: fold the filter+map over vaccinations into one flatMap.
- TaskSlot: 826 lines holding six components. Split so each file declares one:
  TaskSlotGridLines, TaskDropOverlays, TaskMarker, TaskPopoverActionButton and
  TaskDetailsPopover move to their own files, shared marker styling moves to
  taskMarkerStyles.ts, and TaskSlot.tsx drops to 347 lines. The two Storybook
  files now import from the component files directly.

No behaviour change. Validated: tsc clean, eslint 0 errors, 927 tests across
63 suites pass. React Doctor now reports 100/100 for the Modal, petPassport,
passport-share and Calendar/Task directories, which between them held all ten
warnings.
